move validateWebSite to root
